Update Enemy behaviour
Merapihkan Script dari Enemy Controller
