+++ Assets/Prefabs/EnemyBehaviour/Pre_Enemy.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 1657687681640978611}
   - component: {fileID: 1988670560202616659}
   - component: {fileID: 391569849032127248}
-  - component: {fileID: 2800586064894039426}
   - component: {fileID: 4425454995608835221}
   m_Layer: 0
   m_Name: Pre_Enemy
@@ -120,19 +119,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 134d8ad108d4a9a4e8f58cc6a29c7bf4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &2800586064894039426
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8693549782459797256}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5d428f0d031e0b442b8653ace56a928c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speedChase: 5
 --- !u!195 &4425454995608835221
 NavMeshAgent:
   m_ObjectHideFlags: 0
