--- Assets/Scripts/EnemyBehaviour/EnemyAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d428f0d031e0b442b8653ace56a928c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
