--- Assets/Scripts/EnemyBehaviour/EnemySpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c0db0747b5240e74688e7f0a8cb8a646
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
